bugfix: CPP tests use now uint64 symbols

+++ tests/common.cpp
@@ -8,7 +8,7 @@
 #include <random>
 #include <vector>
 
-void fillVectorRandomUniform(uint64_t min, uint64_t max, std::vector<unsigned int> *const vector) {
+void fillVectorRandomUniform(uint64_t min, uint64_t max, std::vector<uint64_t> *const vector) {
     // First create an instance of an engine.
     std::random_device rnd_device;
     // Specify the engine and distribution.
@@ -20,14 +20,14 @@ void fillVectorRandomUniform(uint64_t min, uint64_t max, std::vector<unsigned in
         std::cout << "fillVectorRandomInterval: min: " << int64_t(min) << "; max: " << int64_t(max)
                   << "; seed: " << seed << std::endl;
     }
-    std::uniform_int_distribution<unsigned int> dist(0, max - min);
+    std::uniform_int_distribution<uint64_t> dist(0, max - min);
 
     std::generate(vector->begin(), vector->end(), [&dist, &mersenne_engine, &min]() {
         return dist(mersenne_engine) + min;  // to keep the diff values within the 64-bit signed range
     });
 }
 
-void fillVectorRandomGeometric(std::vector<unsigned int> *const vector) {
+void fillVectorRandomGeometric(std::vector<uint64_t> *const vector) {
     // First create an instance of an engine.
     std::random_device rnd_device;
     // Specify the engine and distribution.
@@ -35,7 +35,7 @@ void fillVectorRandomGeometric(std::vector<unsigned int> *const vector) {
     std::mt19937_64 mersenne_engine{seed};  // Generates random integers
     std::cout << "fillVectorRandomGeometric: len: " << vector->size() << "; seed: " << seed << std::endl;
 
-    std::geometric_distribution<unsigned int> dist(0.05);
+    std::geometric_distribution<uint64_t> dist(0.05);
 
     std::generate(vector->begin(), vector->end(), [&dist, &mersenne_engine]() {
         return dist(mersenne_engine);  // to keep the diff values within the 64-bit signed range
